backend_py/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.1
python-socketio==5.11.4
paho-mqtt==2.1.0
pydantic==2.10.3
python-dotenv==1.0.1
reportlab==4.2.5
numpy==2.1.3
opencv-python-headless==4.10.0.84
ultralytics==8.3.40
